Make qvm-start drive assignment reusable
Qube Manager is calling qvm_start.main on 3 separate locations, and that
function should never have been called from other modules because it's
written to expect a shell client, and exception raised are converted to
what a shell user needs to consume. With the necessity of using
asyncio.run in the main function, it broke Qube Manager as an event loop
was already running, therefore make it use function intended for other
python modules via utils.
